requirements.txt
# RedStudio AI Paint API - Python Dependencies
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# AI/ML
transformers==4.37.0
torch==2.1.2
torchvision==0.16.2

# Utilities
python-dotenv==1.0.0

# Optional (for GPU acceleration)
# torch==2.1.2+cu118
# torchvision==0.16.2+cu118
